refactor(app-cmds): move global checks to Client
